SNN changed into Long

+++ src/main/java/com/example/placebo/controllers/doctors/DoctorResponse.java
@@ -14,7 +14,7 @@ public class DoctorResponse {
     private String title;
     private String name;
     private String surname;
-    private int ssn;
+    private long ssn;
     private String specialisation;
     private int trialId;
